requirements.txt
sqlalchemy
pyyaml
pandas
python-dotenv
aiohttp
aiosqlite  # 如果您使用 SQLite
